Remove confusing constructor (#213)
There is typically no use case for using the binder without defining a
proper local user. Having a constructor that allows not defining a user
invites to a confusing situation where field highlights do not behave as
expected.

If there are cases where it's desirable to not have a proper local user,
then they can pass a dummy value to the remaining constructor.