Add integration tests for production mode (license checking) (#430)
To test the license checking with both a valid license and an invalid license, two new modules have been added. They use Spring, because that's the recommended stack. With the existing Spring IT module, there are now three IT modules that use Spring, which are now grouped under a parent module:

- collaboration-engine-test-spring-parent (common config for all Spring ITs)
  - collaboration-engine-test-spring ("regular" Spring test suite in dev mode)
  - collaboration-engine-test-production (one test case: production mode works with a valid license)
  - collaboration-engine-test-invalid-license (one test case: app breaks with invalid license in production)

Adding two modules that create a production build increases the build time significantly. To avoid the production builds when tests are skipped, all the IT modules are now skipped if -DskipTests is set.

Resolves #408