Clear grace period and events when license file have been changed (#459)
Resolves #441.

License change is based on comparing the license key in ce-license.json and ce-stats.json.
If the license has changed, we now clear grace start date, so that the customer can have a new grace period
We also clear the events GRACE_PERIOD_STARTED, GRACE_PERIOD_ENDED, LICENSE_EXPIRES_SOON, LICENSE_EXPIRED, as a new license would affect all these events.
fireLicenseEvent now also checks that an event not only exists in the list, but that it is also not null. Might be that we clear them for some reason and therefore have a null value.