Always include test-common module in build (#539)
The IT modules are built only when tests are not skipped,
because they're really slow to build.

The test-common module is not as slow, and it's needed by
the other IT modules.

When I run "mvn install" in the project root, I want all
the potential changes I might have done in test-common to
be built. Then it's less likely to have the confusing
situation where you edited some test view in test-common
and then run "mvn verify" in one IT module, accidentally
not including the latest changes in test-common.