Rename Common test bases away from *IT (#541)
File names ending on `*IT` will be automatically run as integration tests (see https://vaadin.com/docs/v14/testbench/testbench-getting-started.html#testbench.quickstart.create-a-test-class).
Because we have base class and extending class, both named *IT (e.g. `AvatarGroupIT`), it means that each test will be ran twice.
This renames the base classes to `*TestCommon` instead of `*IT`, to remove the duplicate runs.

This should make running the test suites a bit faster and avoid having Run 1 / Run 2 in the error logs when a test fails. (Run 2 always failed in a no session exception)