Fix: Move beacon URL to ./?v-r=beacon&id=<uuid> (#604)
Previous URL for the beacon (/beacon/<uuid>)  was being blocked by Spring security's CSRF filtering. That generated an internal redirect to the error page, which could make the error page be saved by the SavedRequestAwareAuthenticationSuccessHandler. This issue was fixed by changing the URL to use to the format used by Flow's internal requests, which should be not be blocked.

Fixes: https://github.com/vaadin/collaboration-engine/issues/41

Co-authored-by: Leif Åstrand <leif@vaadin.com>