feat!: Dispatch asynchronously (#650)
`ComponentConnectionContext` uses an `Executor` to always dispatch tasks asynchronously even in case the target session was currently not locked. This helps resolve deadlocks and will also reduce the performance impact of wide fan-out since the submitting thread doesn't have to wait for all subscriber logic to be run.

These changes did also reveal an API mismatch related to how the connection context could be used to dispatch actions even though it had never been activated. For this reason, the activation handler API was refactored to make the relationship clearer.