fix!: Ignore transient activation or deactivation (#676)
This deals with the situation when the context signals activation before
a pending deactivation has been applied. In that case, the two actions
take out each other so we refrain from running either of them. The same
logic is also applied for the inverse situation with signalling
deactivation before a pending activation is applied.

Co-authored-by: Tulio Garcia <28783969+tulioag@users.noreply.github.com>